estilizado header, a grade de playlist

+++ semana7/labefy/src/components/GradePlayListas/GradePlayLista.jsx
@@ -1,6 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
 import axios from 'axios'
+import backgroundHeader from '../../img/header.jpg'
+import play from '../../img/play.jpg'
+
 import {playlistas, apiUrl, headers} from '../../config/configApi'
 import ListaDeMusicas from '../ListaDeMusicas/ListaDeMusicas'
 
@@ -44,6 +47,7 @@ class GradePlayLista extends React.Component {
         try {
             await axios.post(apiUrl, novaPlayLista, headers)
             this.carregarPlaylistas()
+            alert('Criada com Sucesso!')
             
         } catch (error) {
             console.log('Erro Criar Play lista', error)
@@ -52,13 +56,17 @@ class GradePlayLista extends React.Component {
 
 
       apagarPlaylista = async (id) => {
-          try {
-            await axios.delete(`${apiUrl}/${id}`, headers)
-            this.carregarPlaylistas()
-            alert('Deletada')
-          } catch (error) {
-            console.log('Erro ao Apagar playlista', error)
-          }
+        if(window.confirm('Deseja excluir playlist?')){
+
+            try {
+              await axios.delete(`${apiUrl}/${id}`, headers)
+              this.carregarPlaylistas()
+              alert('Deletada')
+            } catch (error) {
+              console.log('Erro ao Apagar playlista', error)
+            }
+        }
+          
       }
 
 
@@ -66,7 +74,7 @@ class GradePlayLista extends React.Component {
 
 
       verMusicas= (id, nome) => {
-        this.setState({idPlayList: id, mostraMusicas: !this.state.mostraMusicas, nomePlaylist: nome})
+         this.setState({idPlayList: id, mostraMusicas: !this.state.mostraMusicas, nomePlaylist: nome})
       }
 
     render(){
@@ -77,7 +85,7 @@ class GradePlayLista extends React.Component {
 
         return(
             <ContainerPlaylista>
-                <Header>
+                <Header imagem={backgroundHeader}>
                     <h1>Crie sua Playlist</h1>
                     <div>
                         <input placeholder={'Nome'} onChange={this.InputCriarPlaylista}/>
@@ -86,17 +94,19 @@ class GradePlayLista extends React.Component {
                 </Header>
                 <ConteinerConteudo>
                     <Grid>
-                        {this.state.playlistas.map((lista)=>{
+                        {this.state.playlistas.map((lista, indice)=>{
                             return (
-                            <CardLista key={lista.id}>
-                                <p>{lista.name}</p>
-                                <button onClick={()=>this.verMusicas(lista.id, lista.name)}>Ver</button>
-                                <button onClick={()=>this.apagarPlaylista(lista.id)}>Excluir</button>
+                            <CardLista key={lista.id} cardFundo={`https://picsum.photos/id/1${indice}/200/300/?blur=3`}>
+                                <h2>{lista.name}</h2>
+                                <div>
+                                    <BntVer onClick={()=>this.verMusicas(lista.id, lista.name)}>Ver</BntVer>
+                                    <BntExcluir onClick={()=>this.apagarPlaylista(lista.id)}>Excluir</BntExcluir>
+                                </div>
                             </CardLista>
                             )
                         })}
                     </Grid>
-                    <Musicas visibilidade={this.state.mostraMusicas}>
+                    <Musicas visibilidade={this.state.mostraMusicas} fundo={play}>
                          {this.state.mostraMusicas && musicas()}
                     </Musicas>
                 </ConteinerConteudo>
@@ -116,21 +126,24 @@ const ContainerPlaylista = styled.div`
 `
 
 const Header = styled.header`
-    background-color:#f8f8f8;
+    background: url(${props => props.imagem});
     display: flex;
-    height: 13%;
-    padding: 5px;
-    justify-content: flex-start;
+    flex-direction: column;
+    height: 40%;
+    padding-bottom: 1%;
+    justify-content: flex-end;
     align-items: flex-end;
 
     h1{
-       color:#7b7b7b;
-       font-size: 41px;
-       margin: 0;
+       color:#fff;
+       font-size: 50px;
+       line-height: 60px;
+       margin:0 10% 1% 0;
     }
 
     div{
-        margin-left: 20%;
+        margin-right: 12%;
+        
 
         input{
             height: 30px;
@@ -147,8 +160,10 @@ const Header = styled.header`
             height: 32px;
             width: 65px;
             margin-left: 4px;
+            color: #fff;
+            font-size: 17px;
             border: none;
-            background-color: #c5c5c5;
+            background-color: #09f77c;
             box-shadow: 0px 0px 1px 0px black;
             outline: none;
             :hover {
@@ -164,14 +179,14 @@ const Header = styled.header`
 
 const ConteinerConteudo = styled.div`
 display: flex;
-width: 90%;
-height: 80%;
+width: 90vw;
+height: 100vh;
 margin: 10px auto;
 /* border: 1px solid black; */
 /* background-color: #55fbe5; */
 align-items:center;
-justify-content:center;
-overflow-y: scroll;
+justify-content:space-between;
+/* overflow-y: scroll; */
 `
 
 const Grid = styled.div`
@@ -182,8 +197,8 @@ flex-wrap: wrap;
 
 `
 const Musicas = styled.div`
- background-color: #f6ff9d;
-width: 100%;
+background: url(${props => props.fundo});
+width: 80%;
 height: 100%;
 margin: 0;
 display: ${props => props.visibilidade? 'block': 'none'};
@@ -194,5 +209,50 @@ const CardLista = styled.div`
 width: 190px;
 height: 200px;
 margin: 10px 0 0 10px;
-border: 1px solid black;
+background: url(${props => props.cardFundo});
+box-shadow: -1px 3px 8px 0px #ffc5d0;
+display: flex;
+flex-direction: column;
+align-items: center;
+justify-content: space-between;
+
+h2 {
+    font-size: 28px;
+    box-shadow: 1px 1px 1px 0px #54f52a;
+    color: #fff;
+    width: 80%;
+    text-align: center;
+    margin:10px 0 0;
+}
+
+div{
+   display: flex;
+   flex-direction: column;
+   width: 100%;
+   align-items: center;
+}
+
+`
+const BntVer = styled.button`
+    width: 99%;
+    padding: 10px 0;
+    border: none;
+    color: #fff;
+    font-size: 18px;
+    background-color: #09f77c;
+    box-shadow: 0px 0px 1px 0px black;
+    outline: none;
+    :hover {
+        font-size: 19px;
+        box-shadow: 0px 0px 4px 0px black;
+    }
+` 
+const BntExcluir = styled(BntVer)`
+ color: #a09e9e;
+ background-color: #ffffff;
+ font-size: 17px;
+ :hover {
+        color: #ff7b7b;
+        font-size: 18px;
+    }
 `
